feat: Add support for iOS bundle identifier and EAS configuration
This commit adds support for the iOS bundle identifier in the app.json file, setting it to "ios-tournify-bundler". It also includes the EAS configuration in the app.json file, with the project ID set to "a7a81e29-ec87-466e-89b7-2ec6f8aafd1c". Additionally, a new eas.json file is added, which contains configuration settings for EAS build and submission.